Source code in "ref" is in one file
